refactor(VObject): 统一使用VObject变量系统
- 所有变量（variables, magicVariables, namedVariables）统一存储为 VObject
- 新增 ExecutionContext 变量获取便捷方法：getVariableAsString/Number/Boolean 等
- 新增 VEvent 类型支持 UI 事件，统一使用 PropertyRegistry 管理属性
- 优化 LuaValueConverter：VObject 拆箱、MapProxy 零拷贝
- 优化 ConditionEvaluator：支持 Int/Long/Float/Double 严格类型比较
- 新增 ConditionEvaluatorTest 完整单元测试覆盖所有操作符

+++ app/src/main/java/com/chaomixian/vflow/core/execution/LuaValueConverter.kt
@@ -34,7 +34,7 @@ object LuaValueConverter {
 
             // --- VObject 类型拆箱 ---
             is VString -> LuaValue.valueOf(value.raw)
-            is VNumber -> LuaValue.valueOf(value.raw)
+            is VNumber -> LuaValue.valueOf(value.raw.toDouble())
             is VBoolean -> LuaValue.valueOf(value.raw)
             // 列表和字典递归调用自身进行转换
             is VDictionary -> coerceToLua(value.raw)
